move MonitorRunner.py to bin/
