+++ SniperUtil/SniperProfiling/python/SniperMonitor/MonitorRunner.py
@@ -1,10 +1,13 @@
 import sys
+sys.path.append('../python/')
+
 from PySide6.QtWidgets import QApplication
-from SniperMonitor import SniperMonitor
+from SniperMonitor.MainMonitor import MainMonitor
+#from SniperMonitor.MainMonitor import MainMonitor
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
-    sniper_monitor = SniperMonitor()
+    sniper_monitor = MainMonitor()
     sniper_monitor.resize(1000, 500)
     sniper_monitor.show()
     sniper_monitor.AddInfo()
